requirements.txt
cython
h5py
tqdm
msgpack
pillow
imageio
numpy
scipy
scikit-image
gdown
